accuracy up to 0.01

+++ CPU/CPU.h
@@ -2,7 +2,7 @@
 
 struct processor
 {
-    int    registers[4];     //four registers ax, bx, cx, dx
+    double registers[4];     //four registers ax, bx, cx, dx
     struct stack_t box;      // stack with data
     struct stack_t func_ret; //stack with addreses of returning
 };
@@ -34,9 +34,9 @@ enum code_of_com
 
 //--------------------------------------------------------------------------------------
 
-void dist_task    (stack_t* box, stack_t* func_ret, processor* cpu, int elem);
+void dist_task    (stack_t* box, stack_t* func_ret, processor* cpu, double elem);
 int* read_num_com ();
-void stack_add    (stack_t* box, int elem);
-void stack_mul    (stack_t* box, int elem);
-void stack_div    (stack_t* box, int elem);
-void find_sqrt    (stack_t* box, int elem);
+void stack_add    (stack_t* box, double elem);
+void stack_mul    (stack_t* box, double elem);
+void stack_div    (stack_t* box, double elem);
+void find_sqrt    (stack_t* box, double elem);
